Translator: - Migrated to double from long Parser: - Simplified ShuntingYardSyntaxAnalyzer implementation and made it stateless API: - Added endpoint to call evaluation

+++ translator/src/main/java/edu/nix/math/expression/translator/Translator.java
@@ -3,5 +3,5 @@ package edu.nix.math.expression.translator;
 import edu.nix.math.expression.parser.domain.SyntaxTree;
 
 public interface Translator {
-	long translate(SyntaxTree syntaxTree);
+	double translate(SyntaxTree syntaxTree);
 }
